extract files utily functions
